added inserting list items from the toolbox
