+++ src/index.html
@@ -43,7 +43,7 @@
 
                     <div id="add-list-literal-btn" class="button">[]</div>
                     <div id="add-list-item-btn" class="button">
-                        <hole></hole>
+                        , <hole></hole>
                     </div>
                     <div id="add-list-append-stmt-btn" class="button">
                         <span class="hole-toolbox"></span>.append(<hole></hole>)
